adds blur to FAB + refactoring
